update schema/xslt to creat table of anglicisms
This contains xslt to create tables which reflect all of the information
contained in the anglicism tabs. This necessitated the creation of a bf
attribute (bare or base form), since I don't see the need to
differentiate between plurals and singulars. I need to discuss with
someone who understands xslt better about how to compile all of the
anglicisms from all our articles into one table and eliminate multiple
entries while still keeping a count of the occurrence of each distinct
base form